+++ R/glosses_df.R
@@ -7,7 +7,7 @@
 #' @format A data frame with 1339 rows and 4 variables:
 #' \describe{
 #'   \item{gloss}{the gloss abbreviation}
-#'   \item{definition}{the gloss definition}
+#'   \item{definition_en}{the gloss definition}
 #'   \item{source}{the gloss source. Three possible values: Leipzig Glossing Rules, \href{https://en.wikipedia.org/wiki/List_of_glossing_abbreviations}{Wikipedia} or lingglosses (this means parsed from Glossa).}
 #'   \item{weight}{glossa weight used for the choice in case of multiple definitions per gloss.}
 #' }